+++ backend/src/main/java/woowacourse/touroot/travelogue/domain/TravelogueDay.java
@@ -1,4 +1,4 @@
-package woowacourse.touroot.travelogue.domain.photo.domain;
+package woowacourse.touroot.travelogue.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -13,29 +13,25 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import woowacourse.touroot.entity.BaseEntity;
-import woowacourse.touroot.travelogue.domain.place.domain.TraveloguePlace;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Entity
-public class TraveloguePhoto extends BaseEntity {
+public class TravelogueDay extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "PHOTO_KEY", nullable = false)
-    private String key;
-
-    @Column(name = "PHOTO_ORDER", nullable = false)
+    @Column(name = "DAY_ORDER", nullable = false)
     private Integer order;
 
     @JoinColumn(nullable = false)
     @ManyToOne(fetch = FetchType.LAZY)
-    private TraveloguePlace traveloguePlace;
+    private Travelogue travelogue;
 
-    public TraveloguePhoto(Integer order, String key, TraveloguePlace traveloguePlace) {
-        this(null, key, order, traveloguePlace);
+    public TravelogueDay(Integer order, Travelogue travelogue) {
+        this(null, order, travelogue);
     }
 }
