+++ backend/src/main/java/woowacourse/touroot/travelogue/dto/request/TravelogueRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.touroot.travelogue.dto;
+package woowacourse.touroot.travelogue.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
