+++ backend/src/main/java/woowacourse/touroot/travelogue/dto/response/TraveloguePlaceResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.touroot.travelogue.domain.place.dto;
+package woowacourse.touroot.travelogue.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
