+++ backend/src/main/java/woowacourse/touroot/travelogue/dto/response/TravelogueResponse.java
@@ -1,11 +1,10 @@
-package woowacourse.touroot.travelogue.dto;
+package woowacourse.touroot.travelogue.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
 import lombok.Builder;
-import woowacourse.touroot.travelogue.domain.day.dto.TravelogueDayResponse;
 
 @Builder
 public record TravelogueResponse(
