+++ backend/src/main/java/woowacourse/touroot/travelogue/repository/TravelogueDayRepository.java
@@ -1,11 +1,11 @@
-package woowacourse.touroot.travelogue.domain.day.repository;
+package woowacourse.touroot.travelogue.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import woowacourse.touroot.travelogue.domain.Travelogue;
-import woowacourse.touroot.travelogue.domain.day.domain.TravelogueDay;
+import woowacourse.touroot.travelogue.domain.TravelogueDay;
 
 public interface TravelogueDayRepository extends JpaRepository<TravelogueDay, Long> {
-    
+
     List<TravelogueDay> findByTravelogue(Travelogue travelogue);
 }
