+++ backend/src/main/java/woowacourse/touroot/travelogue/repository/TraveloguePhotoRepository.java
@@ -1,9 +1,9 @@
-package woowacourse.touroot.travelogue.domain.photo.repository;
+package woowacourse.touroot.travelogue.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
-import woowacourse.touroot.travelogue.domain.photo.domain.TraveloguePhoto;
-import woowacourse.touroot.travelogue.domain.place.domain.TraveloguePlace;
+import woowacourse.touroot.travelogue.domain.TraveloguePhoto;
+import woowacourse.touroot.travelogue.domain.TraveloguePlace;
 
 public interface TraveloguePhotoRepository extends JpaRepository<TraveloguePhoto, Long> {
