+++ backend/src/main/java/woowacourse/touroot/travelogue/repository/TraveloguePlaceRepository.java
@@ -1,11 +1,11 @@
-package woowacourse.touroot.travelogue.domain.place.repsitory;
+package woowacourse.touroot.travelogue.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
-import woowacourse.touroot.travelogue.domain.day.domain.TravelogueDay;
-import woowacourse.touroot.travelogue.domain.place.domain.TraveloguePlace;
+import woowacourse.touroot.travelogue.domain.TravelogueDay;
+import woowacourse.touroot.travelogue.domain.TraveloguePlace;
 
 public interface TraveloguePlaceRepository extends JpaRepository<TraveloguePlace, Long> {
-    
+
     List<TraveloguePlace> findByTravelogueDay(TravelogueDay travelogueDay);
 }
